Move ContentNodeOptions to components
