+++ contentcuration/contentcuration/frontend/channelEdit/components/ContentNodeOptions.vue
